refactor: 폴더이름과 export폴더 이름 통일
